# [k8s] Remove SSH jump pod for port-forward mode