--- a/sky/provision/kubernetes/config.py
+++ b/sky/provision/kubernetes/config.py
@@ -9,8 +9,8 @@ import yaml
 
 from sky.adaptors import kubernetes
 from sky.provision import common
-from sky.provision.kubernetes import utils as kubernetes_utils
 from sky.provision.kubernetes import network_utils
+from sky.provision.kubernetes import utils as kubernetes_utils
 from sky.utils import kubernetes_enums
 
 logger = logging.getLogger(__name__)
@@ -27,7 +27,8 @@ def bootstrap_instances(
 
     _configure_services(namespace, config.provider_config)
 
-    networking_mode = network_utils.get_networking_mode(config.provider_config.get('networking_mode'))
+    networking_mode = network_utils.get_networking_mode(
+        config.provider_config.get('networking_mode'))
     if networking_mode == kubernetes_enums.KubernetesNetworkingMode.NODEPORT:
         config = _configure_ssh_jump(namespace, config)