--- a/sky/provision/kubernetes/network_utils.py
+++ b/sky/provision/kubernetes/network_utils.py
@@ -44,13 +44,15 @@ def get_port_mode(
 
 
 def get_networking_mode(
-        mode_str: Optional[str] = None) -> kubernetes_enums.KubernetesNetworkingMode:
+    mode_str: Optional[str] = None
+) -> kubernetes_enums.KubernetesNetworkingMode:
     """Get the networking mode from the provider config."""
     mode_str = mode_str or skypilot_config.get_nested(
         ('kubernetes', 'networking_mode'),
         kubernetes_enums.KubernetesNetworkingMode.PORTFORWARD.value)
     try:
-        networking_mode = kubernetes_enums.KubernetesNetworkingMode.from_str(mode_str)
+        networking_mode = kubernetes_enums.KubernetesNetworkingMode.from_str(
+            mode_str)
     except ValueError as e:
         # Add message saying "Please check: ~/.sky/config.yaml" to the error
         # message.
@@ -58,7 +60,6 @@ def get_networking_mode(
             raise ValueError(str(e) + ' Please check: ~/.sky/config.yaml.') \
                 from None
     return networking_mode
-
 
 
 def fill_loadbalancer_template(namespace: str, service_name: str,