Closes #3566. SSH jump pod is not required when using port-forward mode. This PR directly `kubectl port-forwards` to the relevant pod. 

Reduces SSH time by ~25% (1s) on a GKE cluster.

Also lays the groundwork for easy switching between kubecontexts/kubeconfigs while retaining SSH functionality (requested by user).

TODOs - 
- [ ] Need to figure out multi-node support with this, since each worker will now require a unique SSH ProxyCommand. Current SkyPilot plumbing assumes a common ProxyCommand across all workers.


### Benchmarks
======= This branch - After removing SSH Jump pod =======
```
1: ssh test ls
            Mean        Std.Dev.    Min         Median      Max
real        3.014       0.277       2.525       3.125       3.309
user        0.020       0.002       0.016       0.020       0.022
sys         0.008       0.001       0.005       0.008       0.009

multitime -n 5 sky launch -y --cpus 1 --cloud kubernetes task4.yaml
1: sky launch -y --cpus 1 --cloud kubernetes task4.yaml
            Mean        Std.Dev.    Min         Median      Max
real        60.359      2.530       57.249      61.003      63.215
user        4.299       0.492       3.875       3.942       5.095
sys         3.770       0.325       3.492       3.563       4.360
```

======= Master - with SSH Jump pod =======
```
===> multitime results
1: ssh test ls
            Mean        Std.Dev.    Min         Median      Max
real        3.960       0.386       3.310       4.099       4.454
user        0.019       0.002       0.017       0.018       0.022
sys         0.007       0.000       0.006       0.007       0.008

multitime -n 5 sky launch -y --cpus 1 --cloud kubernetes task4.yaml
1: sky launch -y --cpus 1 --cloud kubernetes task4.yaml
            Mean        Std.Dev.    Min         Median      Max
real        63.814      2.132       61.270      63.385      66.609
user        3.791       0.155       3.535       3.854       3.976
sys         3.590       0.062       3.509       3.573       3.670
```

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [ ] Smoke tests